Exercises 1,2 modified

+++ web-ohjelmointi/harj2/tehtava1/script.js
@@ -1,6 +1,6 @@
 function addition() {
-  var a = parseInt(document.getElementById('number1').value);
-  var b = parseInt(document.getElementById('number2').value);
+  var a = parseFloat(document.getElementById('number1').value);
+  var b = parseFloat(document.getElementById('number2').value);
   // if (isNaN(a) === true) a = 0;
   // if (isNaN(b) === true) b = 0;
   var c = a + b;
